Only regenerate keypath if not exists

+++ scripts/prep-test-ssh.sh
@@ -3,9 +3,13 @@ set -e
 SSH_KEY_NAME=${SSH_KEY_NAME:-ci-test-key}
 SSH_KEY_PATH=$(pwd)/ssh-key/ci
 SSH_USER=ci-dev
-rm -rf $SSH_KEY_PATH
-mkdir -p $SSH_KEY_PATH
-ssh-keygen -f "$SSH_KEY_PATH/${SSH_KEY_NAME}" -N ""
+if [ -f "${SSH_KEY_PATH}/${SSH_KEY_NAME}" ]; then
+      echo "${SSH_KEY_PATH}/${SSH_KEY_NAME} exists."
+else
+  rm -rf $SSH_KEY_PATH
+  mkdir -p $SSH_KEY_PATH
+  ssh-keygen -f "$SSH_KEY_PATH/${SSH_KEY_NAME}" -N ""
+fi
 docker kill saido-linux-sshserver | true
 docker rm saido-linux-sshserver || echo 'could not remove container'
 docker run -d -p 2222:2222 -e USER_NAME=$SSH_USER --name saido-linux-sshserver -v ${SSH_KEY_PATH}/${SSH_KEY_NAME}.pub:/config/.ssh/authorized_keys linuxserver/openssh-server
